Use local storage for sessions
