+++ src/hooks/useDarkMode.tsx
@@ -1,29 +1,5 @@
 import { useEffect, useState } from 'react'
-
-const useLocalStorage = (key: string, initialValue?: any) => {
-  const [storedValue, setStoredValue] = useState(() => {
-    try {
-      const item = window.localStorage.getItem(key)
-      return item ? JSON.parse(item) : initialValue
-    } catch (error) {
-      console.warn(error)
-      return initialValue
-    }
-  })
-
-  const setValue = (value: any) => {
-    try {
-      const valueToStore = value instanceof Function ? value(storedValue) : value
-
-      setStoredValue(valueToStore)
-
-      window.localStorage.setItem(key, JSON.stringify(valueToStore))
-    } catch (error) {
-      console.warn(error)
-    }
-  }
-  return [storedValue, setValue]
-}
+import useLocalStorage from './useLocalStorage'
 
 const useDarkMode = () => {
   const [enabled, setEnabled] = useLocalStorage('dark-theme')
