+++ src/hooks/useLocalStorage.ts
@@ -1,0 +1,28 @@
+import { useState } from "react"
+
+const useLocalStorage = (key: string, initialValue?: any) => {
+    const [storedValue, setStoredValue] = useState(() => {
+      try {
+        const item = window.localStorage.getItem(key)
+        return item ? JSON.parse(item) : initialValue
+      } catch (error) {
+        console.warn(error)
+        return initialValue
+      }
+    })
+  
+    const setValue = (value: any) => {
+      try {
+        const valueToStore = value instanceof Function ? value(storedValue) : value
+  
+        setStoredValue(valueToStore)
+  
+        window.localStorage.setItem(key, JSON.stringify(valueToStore))
+      } catch (error) {
+        console.warn(error)
+      }
+    }
+    return [storedValue, setValue]
+  }
+
+export default useLocalStorage;
